Tax Table Migrations Set

+++ database/migrations/2022_11_01_132345_alter_add_payment_to_taxes_table.php
@@ -14,7 +14,7 @@ class AlterAddPaymentToTaxesTable extends Migration
     public function up()
     {
         Schema::table('taxes', function (Blueprint $table) {
-            $table->double('tax_payment')->change();
+            $table->decimal('tax_payment')->change();
         });
     }
